refactor: Amount 클래스 금액 타입 double로 변경

+++ src/main/java/lotto/domain/Amount.java
@@ -4,26 +4,26 @@ import lotto.domain.constants.LottoRule;
 import lotto.message.ErrorMessage;
 
 public class Amount {
-    private final int amount;
+    private final double amount;
 
-    public Amount(int amount) {
+    public Amount(double amount) {
         validate(amount);
         this.amount = amount;
     }
 
-    public void validate(final int amount) {
+    public void validate(final double amount) {
         validateAmountNegativeNumber(amount);
         validateAmountEnough(amount);
         validateAmountDivide(amount);
     }
 
-    public void validateAmountNegativeNumber(final int amount) {
+    public void validateAmountNegativeNumber(final double amount) {
         if (amount < 0) {
             throw new IllegalArgumentException(ErrorMessage.PURCHASE_AMOUNT_IS_NOT_ZERO.getValue());
         }
     }
 
-    public void validateAmountEnough(final int amount) {
+    public void validateAmountEnough(final double amount) {
         if (amount < LottoRule.PRICE.getValue()) {
             throw new IllegalArgumentException(
                     String.format(ErrorMessage.PURCHASE_AMOUNT_IS_NOT_ENOUGH.getValue(),
@@ -32,7 +32,7 @@ public class Amount {
         }
     }
 
-    public void validateAmountDivide(final int amount) {
+    public void validateAmountDivide(final double amount) {
         if (amount % LottoRule.PRICE.getValue() != 0) {
             throw new IllegalArgumentException(
                     String.format(ErrorMessage.PURCHASE_AMOUNT_IS_NOT_THOUSAND_UNITS.getValue(),
@@ -42,6 +42,6 @@ public class Amount {
     }
 
     public int buyCount() {
-        return this.amount / LottoRule.PRICE.getValue();
+        return (int) this.amount / LottoRule.PRICE.getValue();
     }
 }
